refact(mail): rm s3 and separate api module
