--- refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/refarch/email/integration/domain/exception/LoadAttachmentError.java
@@ -1,12 +1,0 @@
-package de.muenchen.refarch.email.integration.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class LoadAttachmentError extends RuntimeException {
-    public LoadAttachmentError(final String message) {
-        super(message);
-    }
-
-    public LoadAttachmentError(final String message, final Exception cause) {
-        super(message, cause);
-    }
-}
