+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/refarch/email/integration/domain/model/Attachment.java
@@ -1,0 +1,8 @@
+package de.muenchen.refarch.email.integration.domain.model;
+
+import jakarta.activation.DataSource;
+
+public record Attachment(
+        String fileName,
+        DataSource file) {
+}
