+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/refarch/email/integration/domain/model/BasicMail.java
@@ -37,5 +37,5 @@ public class BasicMail {
     /**
      * List of attachment paths.
      */
-    private List<String> filePaths;
+    private List<Attachment> attachments;
 }
