+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/test/resources/templates/test-template.ftl
