+++ refarch-integrations/refarch-email-integration/refarch-email-integration-example/src/main/resources/files/test-pdf.pdf
