--- refarch-integrations/refarch-email-integration/refarch-email-integration-java-example/src/main/resources/application-local.yml
@@ -1,19 +1,0 @@
-spring:
-  mail:
-    host: localhost
-    port: 1025
-    username: test@muenchen.de
-    password: secret
-
-server:
-  port: 8087
-
-refarch:
-  mail:
-    from-address: test@muenchen.de
-    default-reply-to-address: no_reply@muenchen.de
-  s3:
-    bucket-name: test-bucket
-    access-key: minio
-    secret-key: Test1234
-    url: http://localhost:9000
