--- refarch-integrations/refarch-email-integration/refarch-email-integration-rest-example/src/main/resources/application-local.yml
@@ -1,32 +1,0 @@
-spring:
-  mail:
-    host: localhost
-    port: 1025
-    username: test@muenchen.de
-    password: secret
-  security:
-    oauth2:
-      client:
-        provider:
-          sso:
-            issuer-uri: http://keycloak:8100/auth/realms/local_realm
-            user-info-uri: ${spring.security.oauth2.client.provider.sso.issuer-uri}/protocol/openid-connect/userinfo
-            jwk-set-uri: ${spring.security.oauth2.client.provider.sso.issuer-uri}/protocol/openid-connect/certs
-        registration:
-          s3:
-            provider: sso
-            authorization-grant-type: client_credentials
-            client-id: local
-            client-secret: client_secret
-
-server:
-  port: 8087
-
-refarch:
-  mail:
-    from-address: test@muenchen.de
-    default-reply-to-address: no_reply@muenchen.de
-  s3:
-    client:
-      document-storage-url: http://localhost:8086
-      enable-security: true
