--- refarch-integrations/refarch-email-integration/refarch-email/pom.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>de.muenchen.refarch</groupId>
-        <artifactId>refarch-email-integration</artifactId>
-        <version>1.6.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>refarch-email</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>refarch-email-api</module>
-        <module>refarch-email-starter</module>
-    </modules>
-
-</project>
