--- refarch-integrations/refarch-email-integration/refarch-email/refarch-email-api/src/main/java/de/muenchen/refarch/email/model/FileAttachment.java
@@ -1,8 +1,0 @@
-package de.muenchen.refarch.email.model;
-
-import jakarta.mail.util.ByteArrayDataSource;
-
-public record FileAttachment(
-        String fileName,
-        ByteArrayDataSource file) {
-}
